requirements.txt
clickhouse-connect>=0.7.16
python-dotenv>=1.0.1
typer>=0.12.3
rich>=13.9.4
requests>=2.31.0

